% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation-methods.R
\name{annotation}
\alias{annotation}
\title{annotation Slot Getter}
\usage{
annotation(object, ...)
}
\arguments{
\item{object}{A \code{ToxicoSet}}

\item{...}{A \code{list} to allow definition of new parameters on this generic}
}
\value{
A \code{list} of named annotaiton
}
\description{
annotation Slot Getter
}
\examples{
data(TGGATESsmall)
annotation(TGGATESsmall)

}
