% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{as,SummarizedExperiment-method}
\alias{as,SummarizedExperiment-method}
\title{Coerce a SummarizedExperiment object to an ExpressionSet}
\usage{
\S4method{as}{SummarizedExperiment}(object, value)
}
\arguments{
\item{object}{An ExpressionSet to coerce to a SummarizedExperiment}

\item{value}{A \code{character} vector specifying the type to coerce to}
}
\value{
The object coerced to the specified type
}
\description{
Coerce a SummarizedExperiment object to an ExpressionSet
}
\section{Warning}{

This method assumes that all slots not present in a SummarizedExperiment
  were moved into the metadata list from the original ExpressionSet
}

