% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{as}
\alias{as}
\title{Define an S4 Generic for the methods::as function}
\usage{
as(object, ..., value)
}
\arguments{
\item{object}{An object to coerce to a different class}

\item{...}{To allow new parameters for this generc}

\item{value}{The type to coerce the object to}
}
\value{
The object updated to the new type
}
\description{
This will allow creation of new definitons for object conversions
}
