% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTSet.R
\name{availableTSets}
\alias{availableTSets}
\title{Return a table of ToxicoSets available for download}
\usage{
availableTSets(canonical = TRUE)
}
\arguments{
\item{canonical}{[`logical`] Should available TSets show only official TSets, or should
user generated TSets be included?}
}
\value{
A data.frame with details about the available ToxicoSet objects
}
\description{
The function fetches a table of all ToxicoSets available for download from
the ToxicoGx server. The table includes the names of the ToxicoSet, the
types of data available in the object, and the date of last update.
}
\details{
Much more information on the processing of the data and data provenance can be found at:
www.orcestra.ca
}
\examples{
if (interactive()){
availablePSets()
}

}
