% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{checkRSetStructure}
\alias{checkRSetStructure}
\title{A function to verify the structure of a RadioSet}
\usage{
checkRSetStructure(rSet, plotDist = FALSE, result.dir = tempdir())
}
\arguments{
\item{rSet}{A \code{RadiOSet} object}

\item{plotDist}{Should the function also plot the distribution of molecular data?}

\item{result.dir}{The path to the directory for saving the plots as a string, defaults to `tempdir()``}
}
\value{
Prints out messages whenever describing the errors found in the structure of the pset object passed in.
}
\description{
This function checks the structure of a PharamcoSet, ensuring that the
correct annotations are in place and all the required slots are filled so
that matching of cells and radiations can be properly done across different types
of data and with other studies.
}
\examples{

checkRSetStructure(Cleveland_small)

}
