% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{Cleveland_small}
\alias{Cleveland_small}
\title{Cleaveland_mut RadioSet subsetted}
\format{RadioSet object}
\usage{
data(Cleveland_small)
}
\description{
Documentation for this dataset will be added at a later date. For now I just
need this package to pass the CRAN checks! This dataset powers the exampe
usage in the roxygen2 documentation for CoreGx.
}
\references{
Lamb et al. The Connectivity Map: using gene-expression signatures to connect small molecules, genes, and disease. Science, 2006.
}
\keyword{datasets}
