% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allGenerics.R
\name{computeLimmaDiffExpr}
\alias{computeLimmaDiffExpr}
\title{Generic method for performing differential expression analysis on an S4 object
  using the limma package}
\usage{
computeLimmaDiffExpr(object, ...)
}
\arguments{
\item{object}{[`S4`] An S4 object to conduct differential expression analysis
on.}

\item{...}{Allow new parameters to be added to this generic.}
}
\value{
To be defined by the method implementation.
}
\description{
Generic method for performing differential expression analysis on an S4 object
  using the limma package
}
