% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-curation.R
\name{curation<-}
\alias{curation<-}
\title{curation<- Slot Setter}
\usage{
curation(object, ...) <- value
}
\arguments{
\item{object}{A \code{ToxicoSet}}

\item{...}{A \code{list} to allow definition of new parameters on this generic}

\item{value}{A \code{list} of curations for the cell and tissues types in the
tSet object}
}
\value{
A copy of the \code{ToxicoSet} with the updated curation slot
}
\description{
curation<- Slot Setter
}
\examples{
data(TGGATESsmall)
curation(TGGATESsmall) <- curation(TGGATESsmall)

}
