% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{datasetType}
\alias{datasetType}
\title{datasetType Generic}
\usage{
datasetType(object, ...)
}
\arguments{
\item{object}{A \code{ToxicoSet} from which to retrieve the dataset type}

\item{...}{A \code{list} containing fall through arguments; this allows
addition of new parameters to methods for this generic}
}
\value{
A \code{character} vector containing the dataset type
}
\description{
A generic for retrieving the dataset type of an tSet object
}
