% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\docType{methods}
\name{dim,RadioSet-method}
\alias{dim,RadioSet-method}
\title{Get the dimensions of a RadioSet}
\usage{
\S4method{dim}{RadioSet}(x)
}
\arguments{
\item{x}{RadioSet}
}
\value{
A named vector with the number of Cells and Drugs in the RadioSet
}
\description{
Get the dimensions of a RadioSet
}
