% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\docType{methods}
\name{dim,ToxicoSet-method}
\alias{dim,ToxicoSet-method}
\title{Get the dimensions of a ToxicoSet}
\usage{
\S4method{dim}{ToxicoSet}(x)
}
\arguments{
\item{x}{ToxicoSet}
}
\value{
A named vector with the number of Cells and Drugs in the ToxicoSet
}
\description{
Get the dimensions of a ToxicoSet
}
\examples{
data(TGGATESsmall)
dim(TGGATESsmall)

}
