% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.convertTsetMolecularProfilesToSE}
\alias{.convertTsetMolecularProfilesToSE}
\title{tSet molecularProfiles from eSets to SEs}
\usage{
.convertTsetMolecularProfilesToSE(tSet)
}
\arguments{
\item{tSet}{\code{S4} A ToxicoSet containing molecular data in ExpressionSets}
}
\value{
\code{S4} A ToxicoSet containing molecular data in a SummarizedExperiments
}
\description{
Converts all ExpressionSet objects within the molecularProfiles slot of a
  ToxicoSet to SummarizedExperiments
}
\keyword{internal}
