% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/linearQuadratic.R
\name{.linearQuadratic}
\alias{.linearQuadratic}
\title{Linear quadratic}
\usage{
.linearQuadratic(D, pars, SF_as_log = TRUE)
}
\arguments{
\item{D}{A vector of drug concentrations}

\item{pars}{Parameters (a, b) of the linear model}

\item{SF_as_log}{Boolen indicating whether survival fraction is logged}
}
\description{
Linear quadratic
}
