% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{.validateTsetMolecularProfilesToSEConversion}
\alias{.validateTsetMolecularProfilesToSEConversion}
\title{Validate tSet molecularProfiles Conversion}
\usage{
.validateTsetMolecularProfilesToSEConversion(tSet_old, tSet_new)
}
\arguments{
\item{tSet_old}{\code{S4} a tSet containing molecularProfiles as ExpressionSets}

\item{tSet_new}{\code{S4} a tSet containing molecularProfiles as SummarizedExperiments}
}
\value{
\code{message} Any slots which are not the same
}
\description{
Checks that all the information contained in an ExpressionSet molecularProfile
  was successfully tranferred to the SummarizedExperiment molecularProfile
}
\keyword{internal}
