% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{drugInfo<-}
\alias{drugInfo<-}
\title{drugInfo<- Generic}
\usage{
drugInfo(object) <- value
}
\arguments{
\item{object}{The \code{ToxicoSet} to replace drug info in}

\item{value}{A \code{data.frame} with the new drug annotations}
}
\value{
Updated \code{ToxicoSet}
}
\description{
Generic for drugInfo replace method
}
\examples{
drugInfo(TGGATES_small) <- drugInfo(TGGATES_small)

}
