% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-drugInfo.R
\name{drugInfo}
\alias{drugInfo}
\title{drugInfo Getter}
\usage{
drugInfo(object)
}
\arguments{
\item{object}{A \code{ToxicoSet} object}
}
\value{
a \code{data.frame} with the drug annotations
}
\description{
Get the drug annotations in a ToxicoSet object
}
\examples{
data(TGGATESsmall)
drugInfo <- drugInfo(TGGATESsmall)

}
