% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{drugNames<-}
\alias{drugNames<-}
\title{drugNames<- Generic
A generic for the drugNames replacement method}
\usage{
drugNames(object) <- value
}
\arguments{
\item{object}{A \code{ToxicoSet} object.}

\item{value}{A \code{data.frame} of replacement values.}
}
\value{
Updated \code{ToxicoSet}
}
\description{
drugNames<- Generic
A generic for the drugNames replacement method
}
\examples{
drugNames(TGGATESsmall) <- drugNames(TGGATESsmall)
}
