% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{drugNames}
\alias{drugNames}
\title{drugNames Generic}
\usage{
drugNames(tSet)
}
\arguments{
\item{tSet}{A \code{ToxicoSet} object from which to retrieve the included
drug names}
}
\value{
A vector of the drug names used in the ToxicoSet
}
\description{
A generic for the drugNames method
}
\examples{
data(TGGATESsmall)
drugName <- drugNames(TGGATESsmall)[1:10]

}
