% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugTimeMolProfCurve.R
\name{drugTimeMolProfCurve}
\alias{drugTimeMolProfCurve}
\title{Compares gene expression for a specificed set of features over specific
  drug dosages vs time}
\usage{
drugTimeMolProfCurve(tSet, duration, cellline, mDataTypes,
  features = NULL, dose, drug, plot.type = "Actual",
  summarize.replicates = FALSE, xlim = c(0, 24), ylim = c(0, 15),
  mycol, x.custom.ticks = NULL, title, lwd = 1.5, cex = 0.7,
  cex.main = 1, legend.loc = "topright", verbose = TRUE)
}
\arguments{
\item{tSet}{\code{ToxicoSet} A ToxicoSet to be plotted in this graph. Currently
only a single tSet is supported, passing more may results in errors.}

\item{duration}{\code{character} A vector of durations to include in the plot.}

\item{cellline}{\code{character} A vector of cell lines to include in the plot.}

\item{mDataTypes}{\code{vector} A vector specifying the molecular data types to
include in this plot. Defaults to the first mDataType if not specified.
This release version only accepts one mDataType, more to be added in
forthcoming releases.}

\item{features}{\code{character} A vector of feature names to include in the plot.
Please note that using too many features will have a significant computational
cost and will likely result in a over crowded plot.}

\item{dose}{\code{character} A vector of dose levels to be included in the
plot. Default to include all dose levels available for a drug. Must include
at minimum two dose levels, one of which must be "Control".}

\item{drug}{\code{character} A vector of drugs to be included in this plot. In
this release, only one drug is supported.}

\item{summarize.replicates}{\code{logical} If true will take the average of all
replicates at each time point per gene and duration. This release has not
yet implemented this feature.}

\item{xlim}{\code{numeric} A vector of minimum and maximum values for the x-axis
of the returned plot.}

\item{ylim}{\code{numeric} A vector of minimum and miximum values for the y-axis
of the returned plot.}

\item{mycol}{\code{vector} A vector of length equal to the number of features
argument specifying which RColorBrewer colour to use per feature. Default
colours will be used if this parameter is excluded.}

\item{x.custom.ticks}{\code{vector} A numeric vector of the distance between major
and minor ticks on the x-axis. If excluded ticks appear only where duration
values are specified.}

\item{title}{\code{character} A string containing the desired plot name. If excluded
a title wil be generated automatically.}

\item{lwd}{\code{numeric} The line width to plot width}

\item{cex}{\code{numeric} The cex parameter passed to plot. Controls the size of
plot points and the font size of the legend and defaults to 0.7.}

\item{cex.main}{\code{numeric} The cex.main parameter passed to plot,
controls the size of the titles and defaults to 1.0.}

\item{legend.loc}{\code{character} The location of the legend as passed to the plot()
function from base graphics. Suggested values are either "topleft" or
"topright", with the latter as the default.}

\item{verbose}{\code{boolean} Should warning messages about the data passed
in be printed?}
}
\value{
Plot of the viabilities for each drug vs time of exposure
}
\description{
This function generates a plot visualizing the relationship between gene
  expression, time and dose level for the selected tSet.
}
\examples{

if (interactive()) {
drugTimeMolProfCurve(TGGATESsmall, dose = c("Control", "Low", "Middle"),
  mDataTypes="rna", drug = drugNames(TGGATESsmall)[1],
  duration = c("2", "8", "24"), features = "ENSG00000000003_at")
}

}
