% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{HCC_sig}
\alias{HCC_sig}
\title{HCC_sig dataset}
\format{character}
\usage{
data(HCC_sig)
}
\description{
A dataset cotaining the gene names associated with the HCC geneset signature
}
\keyword{datasets}
