% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\docType{methods}
\name{mDataNames,RadioSet-method}
\alias{mDataNames,RadioSet-method}
\title{mDataNames}
\usage{
\S4method{mDataNames}{RadioSet}(cSet = rSet)
}
\arguments{
\item{cSet}{The parameter}
}
\value{
Vector of names of the molecular data types
}
\description{
Returns the molecular data names for the RadioSet.
}
\examples{
mDataNames(Cleveland_small)

}
