% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-mDataNames.R
\name{mDataNames,ToxicoSet-method}
\alias{mDataNames,ToxicoSet-method}
\title{mDataNames}
\usage{
\S4method{mDataNames}{ToxicoSet}(object)
}
\arguments{
\item{object}{A \code{ToxicoSet} object}
}
\value{
Vector of names of the molecular data types
}
\description{
Returns the names of the molecular data types available in a ToxicoSet
  object
}
\examples{
mDataNames(TGGATESsmall)

}
