% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{name,ToxicoSet-method}
\alias{name,ToxicoSet-method}
\title{name Getter method}
\usage{
\S4method{name}{ToxicoSet}(object)
}
\arguments{
\item{object}{\code{ToxicoSet} A ToxcioSet object}
}
\value{
\code{character} A string of the tSet's name
}
\description{
Retrieves the name of a tSet
}
\examples{
name(TGGATESsmall)

}
