% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{names,ToxicoSet-method}
\alias{names,ToxicoSet-method}
\title{tSet Name}
\usage{
\S4method{names}{ToxicoSet}(x = tSet)
}
\arguments{
\item{x}{[param] The named parameter from the base R names function. For
internal use only.}

\item{tSet}{[ToxicoSet] A ToxcioSet object}
}
\value{
[character] A string of the tSet's name
}
\description{
Retrieves the name of a tSet
}
\examples{
names(TGGATESsmall)

}
