% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paramErrorChecker.R
\name{paramErrorChecker}
\alias{paramErrorChecker}
\title{An Error Checker for Function Parameters}
\usage{
paramErrorChecker(funName, tSet, ...)
}
\arguments{
\item{funName}{[character] A string of the function name. This argument is
used to match the correct parameter checking conditions with each function.}

\item{...}{[pairlist] A list of all parameters passed as arguements to the
function "funName".}
}
\value{
Returns nothing, this function works by side effects only
}
\description{
This function will take in the params of a function as well as its name.
  Error checking will then be conducted on each of it's parameter arguments
  to ensure they meet the input requirements for that function. Descriptive
  errors are returned if the the arguements do not meet the criteria for that
  function.
}
\keyword{internal}
