% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_tSet.R
\name{pharmacoSettoToxicoSet}
\alias{pharmacoSettoToxicoSet}
\title{Coerce pSet to tSet}
\usage{
pharmacoSettoToxicoSet(pSet)
}
\arguments{
\item{pSet}{A \code{PharmacoSet} class object.}
}
\value{
A \code{ToxicoSet} class object containing all data from the pSet.
}
\description{
Forces a \code{PharmacoSet} class objects to be a \code{ToxicoSet} object.
}
