% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{radiationInfo}
\alias{radiationInfo}
\title{radiationInfo Generic}
\usage{
radiationInfo(rSet)
}
\arguments{
\item{rSet}{A \code{RadioSet} object}
}
\value{
a \code{data.frame} with the radiation annotations
}
\description{
Generic for radiationInfo method
}
\examples{
radiationInfo(Cleveland_small)

}
