% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\name{radiationTypes<-}
\alias{radiationTypes<-}
\title{radiationTypes<- Generic}
\usage{
radiationTypes(object) <- value
}
\arguments{
\item{object}{A \code{RadioSet} object to update}

\item{value}{A \code{character} vector of the new radiation names}
}
\value{
Updated \code{RadioSet}
}
\description{
A generic for the radiationTypes replacement method
}
\examples{
radiationTypes(Cleveland_small) <- radiationTypes(Cleveland_small)

}
