% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{sensitivityInfo<-}
\alias{sensitivityInfo<-}
\title{sensitivityInfo<- Generic}
\usage{
sensitivityInfo(object) <- value
}
\value{
Updated \code{ToxicoSet}
}
\description{
A generic for the sensitivityInfo replacement method
}
\examples{
sensitivityInfo(TGGATES_small) <- sensitivityInfo(TGGATES_small)

}
