% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sensitivityRaw.R
\name{sensitivityRaw<-}
\alias{sensitivityRaw<-}
\title{sensitivityRaw<- Replacement Generic}
\usage{
sensitivityRaw(object, ...) <- value
}
\arguments{
\item{object}{A \code{ToxicoSet} to extract the raw sensitivity data from}

\item{...}{A \code{list} to allow new parameters in specific methods}

\item{value}{A \code{array} containing the raw dose and viability data for
the tSet}
}
\value{
A copy of the \code{ToxicoSet} containing the updated sensitivty data
}
\description{
sensitivityRaw<- Replacement Generic
}
\examples{
data(TGATESsmall)
sensitivityRaw(TGGATESsmall) <- sensitivityRaw(TGGATESsmall)

}
