% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{sensitivityRaw}
\alias{sensitivityRaw}
\title{sensitivityRaw Generic}
\usage{
sensitivityRaw(object, ...)
}
\arguments{
\item{object}{A \code{ToxicoSet} to extract the raw sensitivity data from}

\item{...}{A \code{list} to allow new parameters in specific methods}
}
\value{
A \code{array} containing the raw sensitivity data
}
\description{
sensitivityRaw Generic
}
\examples{
data(TGGATESsmall)
sensitivityRaw(TGGATESsmall)

}
