% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RadioSet.R
\docType{methods}
\name{show,RadioSet-method}
\alias{show,RadioSet-method}
\title{Show a RadioSet}
\usage{
\S4method{show}{RadioSet}(object)
}
\arguments{
\item{object}{A \code{RadioSet} object}
}
\value{
Prints the RadioSet object to the output stream, and returns invisible NULL.
}
\description{
Show a RadioSet
}
\examples{
Cleveland_small

}
