% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClass.R
\docType{methods}
\name{show,ToxicoSig-method}
\alias{show,ToxicoSig-method}
\title{Show ToxicoGx Signatures}
\usage{
\S4method{show}{ToxicoSig}(object)
}
\arguments{
\item{object}{\code{ToxicoSig}}
}
\value{
Prints the ToxicoGx Signatures object to the output stream, and returns invisible NULL.
}
\description{
Show ToxicoGx Signatures
}
\examples{
data(TGGATES_small)
drug.perturbation <- drugPerturbationSig(TGGATES_small, mDataType="rna", nthread=1, duration = 2,
     drugs = head(drugNames(TGGATES_small)), features = fNames(TGGATES_small, "rna")[1])
drug.perturbation

}
