% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarizeMolecularProfiles_2.R
\name{summarizeToxicoMolecularProfiles2}
\alias{summarizeToxicoMolecularProfiles2}
\title{Takes molecular data from a ToxicoSet, and summarises them
into one entry per drug for the requested duration}
\usage{
summarizeToxicoMolecularProfiles2(tSet, mDataType, cell.lines, drugs,
  features, duration = c(2, 8, 24), dose = c("Control", "Low",
  "Middle", "High"), summary.stat = c("mean", "median", "first", "last"),
  fill.missing = TRUE, summarize = TRUE, verbose = TRUE)
}
\arguments{
\item{tSet}{\code{ToxicoSet} The ToxicoSet to summarize}

\item{mDataType}{\code{character} which one of the molecular data types
to use in the analysis, out of all the molecular data types available for the tSet
for example: rna}

\item{cell.lines}{\code{character} The cell lines to be summarized.
If any cell.line has no data, missing values will be created}

\item{drugs}{\code{character} The drugs to be summarized}

\item{features}{\code{character} A vector of the feature names to include in the summary}

\item{duration}{\code{numeric} A vector of durations to summarize across}

\item{dose}{\code{character} The dose level to summarize replicates across}

\item{summary.stat}{\code{character} which summary method to use if there are repeated
cell.lines? Choices are "mean", "median", "first", or "last"}

\item{fill.missing}{\code{boolean} should the missing cell lines not in the
molecular data object be filled in with missing values?}

\item{summarize}{A flag which when set to FALSE (defaults to TRUE) disables summarizing and
returns the data unchanged as a ExpressionSet}

\item{verbose}{\code{boolean} should messages be printed}
}
\value{
\code{ExpressionSet} An ExpressionSet with the molecular data summarized
  per cell line.
}
\description{
Given a ToxicoSet with molecular data, this function will summarize
the data into one profile per cell line, using the chosen summary.stat.
}
\examples{
data(TGGATES_small)
TGGATES_small <- summarizeMolecularProfiles2(TGGATES_small,
                    mDataType = "rna", cell.lines=cellNames(TGGATES_small),
                    summary.stat = 'median', fill.missing = TRUE, verbose=TRUE)
TGGATES_small

}
