% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dataset.R
\docType{data}
\name{TGGATESsmall}
\alias{TGGATESsmall}
\title{TGGATESsmall dataset}
\format{ToxicoSet object}
\usage{
data(TGGATESsmall)
}
\description{
Documentation for this dataset will be added at a later date. For now I just
need this package to pass the CRAN checks! This dataset powers the example
usage in the roxygen2 documentation for ToxicoGx.
}
\references{
Lamb et al. The Connectivity Map: using gene-expression signatures to connect
  small molecules, genes, and disease. Science, 2006.
}
\keyword{datasets}
