% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{sensitivitySlot<-,ToxicoSet,list-method}
\alias{sensitivitySlot<-,ToxicoSet,list-method}
\alias{datasetType,ToxicoSet-method}
\title{datasetType Getter}
\usage{
\S4method{sensitivitySlot}{ToxicoSet,list}(object) <- value

\S4method{datasetType}{ToxicoSet}(object)
}
\arguments{
\item{object}{A \code{ToxicoSet} from which to retrieve the dataset type}

\item{value}{A \code{list} of new sensitivity slot data for the tSet}

\item{...}{A \code{list} containing fall through arguments; this allows
addition of new parameters to methods for this generic}
}
\description{
datasetType Getter
}
\section{Functions}{
\itemize{
\item \code{sensitivitySlot<-,ToxicoSet,list-method}: Set the raw dose and viability data for an tSet and return
and updated copy

\item \code{datasetType,ToxicoSet-method}: Update the dataset type of an tSet and return a copy of
the updated object
}}

\examples{
data(TGGATESsmall)
datasetType(TGGATESsmall)

}
