% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-signature.R
\name{ToxicoSig}
\alias{ToxicoSig}
\title{ToxicoSig Constructor}
\usage{
ToxicoSig(
  Data = array(NA, dim = c(0, 0, 0)),
  tSetName = "",
  DateCreated = date(),
  SigType = "sensitivity",
  SessionInfo = sessionInfo(),
  Call = "No Call Recorded",
  Arguments = list()
)
}
\arguments{
\item{Data}{`array`` An array contiaining the data for constructing the ToxicoSig object}

\item{tSetName}{\code{character(1)} The name of the tSet used in the constructor}

\item{DateCreated}{\code{date} The data at time of running the constructor}

\item{SigType}{\code{character}A string of the experiment type}

\item{SessionInfo}{\code{sessionInfo}The current session info}

\item{Call}{\code{character(1)} A string}

\item{Arguments}{\code{list} A list of arguments passed to the constructor}
}
\value{
\code{object} A new ToxicoSig object
}
\description{
A user friendly constructor to create ToxicoSig class objects. This function
is implemented as an internal and should only be called for development purposes
}
\keyword{internal}
