% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{updateDrugId}
\alias{updateDrugId}
\title{A function to update drug ids}
\usage{
updateDrugId(tSet, new.ids = vector("character"))
}
\arguments{
\item{tSet}{[object] A ToxicoSet object to be updates}

\item{new.ids}{[character] A character vector of ids to update with}
}
\description{
A function to update drug ids
}
\examples{
data(TGGATESsmall)
updateDrugId(TGGATESsmall, new.ids = drugNames(TGGATESsmall))
}
\keyword{internal}
