% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/updateObject-methods.R
\name{updateObject,ToxicoSet-method}
\alias{updateObject,ToxicoSet-method}
\title{Update the ToxicoSet class after changes in it struture or API}
\usage{
\S4method{updateObject}{ToxicoSet}(object)
}
\arguments{
\item{object}{A \code{ToxicoSet} object to update the class structure for.}
}
\value{
\code{ToxicoSet} with update class structure.
}
\description{
Update the ToxicoSet class after changes in it struture or API
}
