% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilityFunctions.R
\name{warn_verbose}
\alias{warn_verbose}
\title{Warning if verbose}
\usage{
warn_verbose(warning)
}
\arguments{
\item{warning}{\code{character} A warning message to display if verbose is True}
}
\description{
Warning if verbose
}
\keyword{internal}
