% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{as,DFrame-method}
\alias{as,DFrame-method}
\title{Coerce a DFrame object to an AnnotatedDataFrame}
\usage{
\S4method{as}{DFrame}(object, value)
}
\arguments{
\item{object}{A DFrame object to coerce to AnnotatedDataFrame}

\item{value}{The type to convert to; if not "AnnotatedDataFrame" this
method falls back to as}
}
\description{
Coerce a DFrame object to an AnnotatedDataFrame
}
