% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limmaDE.R
\name{computeLimmaDiffExpr}
\alias{computeLimmaDiffExpr}
\title{Conduct differential expression analysis using the limma R pacakge}
\usage{
computeLimmaDiffExpr(tSet, buildTable = TRUE)
}
\arguments{
\item{tSet}{A [`ToxicoSet`] object with a molecular profile named 'rna'}

\item{buildtable}{[`logical`] Should the result of the eBayes function
from limma be assembled into a data.table containing the result along
with the gene, compound and durations names. Default it TRUE, otherwise
this function with return the object produced by eBayes.}
}
\value{
A [`data.table`] containing the results the limma differential
   expression analysis comparing control vs each dose level for each compound
   within each duration.
}
\description{
Conduct differential expression analysis using the limma R pacakge
}
\section{WARNING}{
 This function can take a very long time to compute!
}

