% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugGeneResponseCurve.R
\name{drugGeneResponseCurve}
\alias{drugGeneResponseCurve}
\title{Compares gene expression for a specificed set of features over specific
  drug dosages vs time}
\usage{
drugGeneResponseCurve(
  tSet,
  duration,
  cell.lines,
  mDataTypes,
  features = NULL,
  dose,
  drug,
  summarize.replicates = TRUE,
  xlim = c(0, 24),
  ylim = c(0, 15),
  mycol,
  title,
  lwd = 1.5,
  cex = 1,
  cex.main = 0.9,
  verbose = TRUE
)
}
\arguments{
\item{tSet}{\code{ToxicoSet} A ToxicoSet to be plotted in this graph. Currently
only a single tSet is supported.}

\item{duration}{\code{character} A vector of durations to include in the plot.}

\item{cell.lines}{\code{character} A vector of cell lines to include in the plot.
Currently limited to one cell lines per plot with plans to add support for
more in upcoming releases.}

\item{mDataTypes}{\code{vector} A vector specifying the molecular data types to
include in this plot. Defaults to the first mDataType if not specified.ex
This release version only accepts one mDataType, more to be added in
forthcoming releases.}

\item{features}{\code{character} A vector of feature names to include in the plot.
If you specify more than two dose levels, you may only pass in up to two features.}

\item{dose}{\code{character} A vector of dose levels to be included in the
plot. Default to include all dose levels available for a drug. If you specify
more than two features you may only pass in up to two dose levels.}

\item{drug}{\code{character} A drug name to include in this plot.
See drugNames(tSet) for a list of options.}

\item{summarize.replicates}{\code{logical} If true will take the average of all
replicates at each time point per gene and duration. This release has not
yet implemented this feature.}

\item{xlim}{\code{numeric} A vector of minimum and maximum values for the x-axis
of the returned plot.}

\item{ylim}{\code{numeric} A vector of minimum and miximum values for the y-axis}

\item{title}{\code{character} A string containing the desired plot name. If excluded
a title wil be generated automatically.
number of drugs, features and doses passed to the function. Takes colour
arguments as passed to `col` parameter in the `plot()` function.
Default palette is used when unspecified.}

\item{lwd}{\code{numeric} The line width to plot width}

\item{cex}{\code{numeric} The cex parameter passed to plot. Controls the size of
plot points and the font size of the legend and defaults to 0.7.}

\item{verbose}{\code{boolean} Should warning messages about the data passed
in be printed?}

\item{trunc}{\code{bool} Should the viability values be truncated to lie in
\code{0-100} before doing the fitting}
}
\value{
Plot of the viabilities for each drug vs time of exposure
}
\description{
This function generates a plot visualizing the relationship between gene
  expression, time and dose level for the selected tSet.
}
\examples{

if (interactive()) {
drugGeneResponseCurve(TGGATESsmall, dose = c("Control", "Low", "Middle"), mDataTypes="rna", drug = drugNames(TGGATESsmall)[1], duration = c("2", "8", "24"), features = "ENSG00000000003_at")
}

}
