% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{drugInfo<-}
\alias{drugInfo<-}
\title{drugInfo<- Generic}
\usage{
drugInfo(object) <- value
}
\arguments{
\item{object}{A \code{ToxicoSet} object.}

\item{value}{A \code{data.frame} of replacement values.}
}
\value{
Updated \code{ToxicoSet}
}
\description{
Generic for drugInfo replace method
}
\examples{
data(TGGATESsmall)
drugInfo(TGGATESsmall) <- drugInfo(TGGATESsmall)

}
