% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{mDataNames,ToxicoSet-method}
\alias{mDataNames,ToxicoSet-method}
\title{mDataNames}
\usage{
\S4method{mDataNames}{ToxicoSet}(cSet = tSet)
}
\arguments{
\item{cSet}{Parameter name for parent method inherited from CoreGx}

\item{tSet}{A \code{ToxicoSet} object}
}
\value{
Vector of names of the molecular data types
}
\description{
Returns the molecular data names for the ToxicoSet.
}
\examples{
mDataNames(TGGATESsmall)

}
