% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods-sensitivitySlot.R
\name{sensitivitySlot<-}
\alias{sensitivitySlot<-}
\title{sensitivitySlot<- Replacement Generic}
\usage{
sensitivitySlot(object, ...) <- value
}
\arguments{
\item{object}{A \code{ToxicoSet} to extract the raw sensitivity data from}

\item{...}{Allow new parameters to be defined for this generic}

\item{value}{A \code{list} of new sensitivity slot data for the tSet}
}
\value{
A copy of the \code{ToxicoSet} containing the updated sensitivty slot
}
\description{
sensitivitySlot<- Replacement Generic
}
