% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{sensitivitySlot}
\alias{sensitivitySlot}
\title{sensitivitySlot Generic}
\usage{
sensitivitySlot(object, ...)
}
\arguments{
\item{object}{A \code{ToxicoSet} to extract the raw sensitivity data from}

\item{...}{Allow new parameters to be defined for this generic}
}
\value{
A \code{list} of the sensitivity slot contents
}
\description{
sensitivitySlot Generic
}
