% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\name{show,ToxicoSet-method}
\alias{show,ToxicoSet-method}
\title{Show a ToxicoSet}
\usage{
\S4method{show}{ToxicoSet}(object)
}
\arguments{
\item{object}{A \code{ToxicoSet} object to print a summary for}
}
\value{
Prints the ToxicoSet object to the output stream, and returns invisible NULL.
}
\description{
Show a ToxicoSet
}
\examples{
TGGATESsmall

}
