% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ToxicoSet.R
\docType{class}
\name{ToxicoSet-class}
\alias{ToxicoSet-class}
\alias{.ToxicoSet}
\alias{cellInfo,ToxicoSet-method}
\alias{cellInfo<-,ToxicoSet,data.frame-method}
\alias{drugInfo,ToxicoSet-method}
\alias{drugInfo<-,ToxicoSet,data.frame-method}
\alias{phenoInfo,ToxicoSet,character-method}
\alias{phenoInfo<-,ToxicoSet,character,data.frame-method}
\alias{molecularProfiles,ToxicoSet,character-method}
\alias{molecularProfiles<-,ToxicoSet,character,matrix-method}
\alias{featureInfo,ToxicoSet,character-method}
\alias{featureInfo<-,ToxicoSet,character,data.frame-method}
\alias{sensitivityInfo,ToxicoSet-method}
\alias{sensitivityInfo<-,ToxicoSet,data.frame-method}
\alias{sensitivityProfiles,ToxicoSet-method}
\alias{sensitivityProfiles<-,ToxicoSet,data.frame-method}
\alias{sensitivityProfiles<-,ToxicoSet,matrix-method}
\alias{sensitivityMeasures,ToxicoSet-method}
\alias{drugNames,ToxicoSet-method}
\alias{drugNames<-,ToxicoSet,character-method}
\alias{cellNames,ToxicoSet-method}
\alias{cellNames<-,ToxicoSet,character-method}
\alias{fNames,ToxicoSet,character-method}
\alias{dateCreated,ToxicoSet-method}
\alias{cSetName,ToxicoSet-method}
\alias{pertNumber,ToxicoSet-method}
\alias{sensNumber,ToxicoSet-method}
\alias{pertNumber<-,ToxicoSet,array-method}
\alias{sensNumber<-,ToxicoSet,matrix-method}
\title{A Class to Contain ToxicoGenomic datasets together with their curations}
\usage{
\S4method{cellInfo}{ToxicoSet}(cSet = tSet)

\S4method{cellInfo}{ToxicoSet,data.frame}(object) <- value

\S4method{drugInfo}{ToxicoSet}(tSet)

\S4method{drugInfo}{ToxicoSet,data.frame}(object) <- value

\S4method{phenoInfo}{ToxicoSet,character}(cSet = tSet, mDataType)

\S4method{phenoInfo}{ToxicoSet,character,data.frame}(object,
  mDataType) <- value

\S4method{molecularProfiles}{ToxicoSet,character}(cSet = tSet, mDataType)

\S4method{molecularProfiles}{ToxicoSet,character,matrix}(object,
  mDataType) <- value

\S4method{featureInfo}{ToxicoSet,character}(cSet = tSet, mDataType)

\S4method{featureInfo}{ToxicoSet,character,data.frame}(object,
  mDataType) <- value

\S4method{sensitivityInfo}{ToxicoSet}(cSet = tSet)

\S4method{sensitivityInfo}{ToxicoSet,data.frame}(object) <- value

\S4method{sensitivityProfiles}{ToxicoSet}(cSet = tSet)

\S4method{sensitivityProfiles}{ToxicoSet,data.frame}(object) <- value

\S4method{sensitivityProfiles}{ToxicoSet,matrix}(object) <- value

\S4method{sensitivityMeasures}{ToxicoSet}(cSet = tSet)

\S4method{drugNames}{ToxicoSet}(tSet)

\S4method{drugNames}{ToxicoSet,character}(object) <- value

\S4method{cellNames}{ToxicoSet}(cSet = tSet)

\S4method{cellNames}{ToxicoSet,character}(object) <- value

\S4method{fNames}{ToxicoSet,character}(cSet = tSet, mDataType)

\S4method{dateCreated}{ToxicoSet}(cSet = tSet)

\S4method{cSetName}{ToxicoSet}(cSet = tSet)

\S4method{pertNumber}{ToxicoSet}(cSet = tSet)

\S4method{sensNumber}{ToxicoSet}(cSet = tSet)

\S4method{pertNumber}{ToxicoSet,array}(object) <- value

\S4method{sensNumber}{ToxicoSet,matrix}(object) <- value
}
\arguments{
\item{cSet}{Parameter name for parent method inherited from CoreGx}

\item{object}{A \code{ToxicoSet} object}

\item{value}{A replacement value}

\item{tSet}{A \code{ToxicoSet} object}

\item{mDataType}{A \code{character} with the type of molecular data to return/update}
}
\value{
An object of the ToxicoSet class

a \code{data.frame} with the cell annotations

Updated \code{ToxicoSet}

a \code{data.frame} with the experiment info

The updated \code{ToxicoSet}

a \code{data.frame} with the experiment info

Updated \code{ToxicoSet}

Updated \code{ToxicoSet}

a \code{data.frame} with the experiment info

a \code{data.frame} with the experiment info

Updated \code{ToxicoSet}

A \code{character} vector of all the available sensitivity measures

A vector of the cell names used in the ToxicoSet

Updated \code{ToxicoSet}

A \code{character} vector of the feature names

The date the ToxicoSet was created

The name of the ToxicoSet

A 3D \code{array} with the number of perturbation experiments per radiation type and cell line, and data type

A \code{data.frame} with the number of sensitivity experiments per drug and cell line

The updated \code{ToxicoSet}

The updated \code{ToxicoSet}
}
\description{
// D E S C R I P T I O N //

Generic for phenoInfo method

Generic for phenoInfo replace method

Generic for molecularProfiles method

Generic for molecularProfiles replace method

Generic for featureInfo replace method

Generic for sensitivityInfo method

Generic for sensitivityProfiles method

A generic for the sensitivityProfiles replacement method

A generic for the sensitivityMeasures  method

A generic for the cellNames method

A generic for the cellNames replacement method

A generic for the fNames method

A generic for the dateCreated method

A generic for the tSetName method

A generic for the pertNumber method

A generic for the sensNumber method

A generic for the pertNumber method

A generic for the sensNumber method
}
\section{Methods (by generic)}{
\itemize{
\item \code{cellInfo}: Returns the annotations for all the cell lines tested on in the ToxicoSet

\item \code{cellInfo<-}: Update the cell line annotations

\item \code{drugInfo}: Returns the annotations for all the drugs tested in the ToxicoSet

\item \code{drugInfo<-}: Update the drug annotations

\item \code{phenoInfo}: Return the experiment info from the given type of molecular data in ToxicoSet

\item \code{phenoInfo<-}: Update the the given type of molecular data experiment info in the ToxicoSet

\item \code{molecularProfiles}: Return the given type of molecular data from the ToxicoSet

\item \code{molecularProfiles<-}: Update the given type of molecular data from the ToxicoSet

\item \code{featureInfo}: Return the feature info for the given molecular data

\item \code{featureInfo<-}: Replace the gene info for the molecular data

\item \code{sensitivityInfo}: Return the drug dose sensitivity experiment info

\item \code{sensitivityInfo<-}: Update the sensitivity experiment info

\item \code{sensitivityProfiles}: Return the phenotypic data for the drug dose sensitivity

\item \code{sensitivityProfiles<-}: Update the phenotypic data for the drug dose
sensitivity

\item \code{sensitivityProfiles<-}: Update the phenotypic data for the drug dose
sensitivity

\item \code{sensitivityMeasures}: Returns the available sensitivity profile
summaries, for example, whether there are IC50 values available

\item \code{drugNames}: Return the names of the drugs used in the ToxicoSet

\item \code{drugNames<-}: Update the drug names used in the dataset

\item \code{cellNames}: Return the cell names used in the dataset

\item \code{cellNames<-}: Update the cell names used in the dataset

\item \code{fNames}: Return the feature names used in the dataset

\item \code{dateCreated}: Return the date the ToxicoSet was created

\item \code{cSetName}: Return the name of the ToxicoSet

\item \code{pertNumber}: Return the summary of available perturbation
experiments

\item \code{sensNumber}: Return the summary of available sensitivity
experiments

\item \code{pertNumber<-}: Update the summary of available perturbation
experiments

\item \code{sensNumber<-}: Update the summary of available sensitivity
experiments
}}

\section{Slots}{

\describe{
\item{\code{annotation}}{A \code{list} of annotation data about the ToxicoSet,
including the \code{$name} and the session information for how the object
was creating, detailing the exact versions of R and all the packages used}

\item{\code{molecularProfiles}}{A \code{list} containing 4 \code{Biobase::ExpressionSet}
type object for holding data for RNA, DNA, SNP and Copy Number Variation
measurements respectively, with associated \code{fData} and \code{pData}
containing the row and column metadata}

\item{\code{cell}}{A \code{data.frame} containg the annotations for all the cell
lines profiled in the data set, across all data types}

\item{\code{drug}}{A \code{data.frame} containg the annotations for all the drugs
profiled in the in the dataset, across all data types}

\item{\code{sensitivity}}{A \code{list} containing all the data for the sensitivity
experiments, including \code{$info}, a \code{data.frame} containing the
experimental info,\code{$raw} a 3D \code{array} containing raw data,
\code{$profiles}, a \code{data.frame} containing sensitivity profiles
statistics, and \code{$n}, a \code{data.frame} detailing the number of
experiments for each cell-drug type pair}

\item{\code{perturbation}}{A \code{list} containting \code{$n}, a \code{data.frame}
summarizing the available perturbation data,}

\item{\code{curation}}{A \code{list} containing mappings for
\code{cell}, \code{drug}, and \code{tissue} names used in the data set to universal
identifiers used between different ToxicoSet objects}

\item{\code{datasetType}}{A \code{character} string of 'sensitivity',
'perturbation', or both detailing what type of data can be found in the
ToxicoSet, for proper processing of the data}
}}

\examples{
data(TGGATES_small)
phenoInfo(TGGATES_small, mDataType="rna")


phenoInfo(TGGATES_small, mDataType="rna") <- phenoInfo(TGGATES_small, mDataType="rna")

TGGATES_mProf <- molecularProfiles(TGGATES_small, "rna")
TGGATES_mProf[1:10,]

molecularProfiles(TGGATES_small, "rna") <- molecularProfiles(TGGATES_small, "rna")

featureInfo(TGGATES_small, "rna")[1:10,]

featureInfo(TGGATES_small, "rna") <- featureInfo(TGGATES_small, "rna")

sensInf<- sensitivityInfo(TGGATES_small)
sensInf[1:10,]

sensitivityProfiles(TGGATES_small)

sensitivityProfiles(TGGATES_small) <- sensitivityProfiles(TGGATES_small)

sensitivityMeasures(TGGATES_small)

cellNames(TGGATES_small)

cellNames(TGGATES_small) <- cellNames(TGGATES_small)

fNames(TGGATES_small, "rna")[1:10]

dateCreated(TGGATES_small)

tSetName <- cSetName
tSetName(TGGATES_small)

pertNumber(TGGATES_small)

sensNumber(TGGATES_small)

pertNumber(TGGATES_small) <- pertNumber(TGGATES_small)

sensNumber(TGGATES_small) <- sensNumber(TGGATES_small)

}
