% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signatureClass.R
\name{ToxicoSig}
\alias{ToxicoSig}
\title{ToxicoSig Constructor}
\usage{
ToxicoSig(
  Data = array(NA, dim = c(0, 0, 0)),
  tSetName = "",
  DateCreated = date(),
  SigType = "sensitivity",
  SessionInfo = sessionInfo(),
  Call = "No Call Recorded",
  Arguments = list()
)
}
\arguments{
\item{Data}{[array] An array contiaining the data for constructing the ToxicoSig object}

\item{tSetName}{[string] The name of the tSet used in the constructor}

\item{DateCreated}{[date] The data at time of running the constructor}

\item{SigType}{[character] A string of the experiment type}

\item{SessionInfo}{[sessionInfo] The current session info}

\item{Call}{[character] A string}

\item{Arguments}{[list] A list of arguments passed to the constructor}
}
\value{
[object] A new ToxicoSig object
}
\description{
A user friendly constructor to create ToxicoSig class objects. This function
  is implemented as an internal and should only be called for development purposes
}
\keyword{internal}
